Make date filter's default year dynamic
This is a partial fix for LOGSTASH-1744. With this change, the date filter will
ensure that the joda time parser is correctly configured with the current year
before parsing the date. This check happens each time a date is parsed.

Even with this change in place, there will still be some bad behavior at the
new year due to clock skew and/or network latency. For example, if a log event
is generated on a host whose clock is slightly ahead (already in the new year),
logstash will incorrectly parse it with the current year instead of the new
year.

Likewise, if a log event is generated on a host whose clock is slightly behind
(still in the previous year), or even if the clocks are perfectly in sync but
logstash's parsing is delayed due to network latency, logstash will incorrectly
parse it with the current year instead of the previous year.